Wrap entry point
